% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_respondents_logits.R
\name{run_respondents_logits}
\alias{run_respondents_logits}
\title{Individual-level models}
\usage{
run_respondents_logits(exercise_row)
}
\arguments{
\item{exercise_row}{NO DEFAULT.}
}
\description{
Runs an individual-level multinomial logit model for each respondent
}
\examples{
# No examples. Good luck
}
